requirements.txt
streamlit==1.41.0
openai-whisper
numpy==1.26.4
google-generativeai==0.8.4
fpdf2==2.8.2
python-dotenv==1.0.1
requests
